resource models re-generated
