+++ pkg/acceptance/bettertestspoc/config/model/secret_with_client_credentials_model_ext.go
@@ -1,0 +1,13 @@
+package model
+
+import tfconfig "github.com/hashicorp/terraform-plugin-testing/config"
+
+func (s *SecretWithClientCredentialsModel) WithOauthScopes(oauthScopes []string) *SecretWithClientCredentialsModel {
+	oauthScopesStringVariables := make([]tfconfig.Variable, len(oauthScopes))
+	for i, v := range oauthScopes {
+		oauthScopesStringVariables[i] = tfconfig.StringVariable(v)
+	}
+
+	s.OauthScopes = tfconfig.SetVariable(oauthScopesStringVariables...)
+	return s
+}
